+++ Sources/Scenes/Stories/StoriesViewController.swift
@@ -6,7 +6,7 @@ class StoriesViewController: UIViewController {
 
     private let feed: APIClient.StoryFeed
 
-    private var segmentedProgress: StoriesProgressBar!
+    private var progress: StoriesProgressBar!
 
     private let menuButton: UIButton = {
         let button = UIButton()
@@ -51,12 +51,12 @@ class StoriesViewController: UIViewController {
     override func viewDidLoad() {
         view.backgroundColor = .black
 
-        segmentedProgress = StoriesProgressBar(numberOfSegments: player.player.items().count)
-        segmentedProgress.translatesAutoresizingMaskIntoConstraints = false
-        segmentedProgress.topColor = UIColor.white
-        segmentedProgress.padding = 5.0
-        segmentedProgress.bottomColor = UIColor.white.withAlphaComponent(0.25)
-        segmentedProgress.dataSource = self
+        progress = StoriesProgressBar(numberOfSegments: feed.stories.count)
+        progress.translatesAutoresizingMaskIntoConstraints = false
+        progress.topColor = UIColor.white
+        progress.padding = 5.0
+        progress.bottomColor = UIColor.white.withAlphaComponent(0.25)
+        progress.dataSource = self
 
         let background = UIView()
         background.translatesAutoresizingMaskIntoConstraints = false
@@ -70,7 +70,7 @@ class StoriesViewController: UIViewController {
         content.backgroundColor = .brandColor
         background.addSubview(content)
 
-        content.addSubview(segmentedProgress)
+        content.addSubview(progress)
 
         let buttonStack = UIStackView()
         buttonStack.translatesAutoresizingMaskIntoConstraints = false
@@ -123,6 +123,32 @@ class StoriesViewController: UIViewController {
         fire.addTarget(self, action: #selector(didReact), for: .touchUpInside)
         heart.addTarget(self, action: #selector(didReact), for: .touchUpInside)
 
+        let rightTapView = UIView()
+        rightTapView.translatesAutoresizingMaskIntoConstraints = false
+        rightTapView.isUserInteractionEnabled = true
+        rightTapView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(skip)))
+        content.addSubview(rightTapView)
+
+        let leftTapView = UIView()
+        leftTapView.translatesAutoresizingMaskIntoConstraints = false
+        leftTapView.isUserInteractionEnabled = true
+        leftTapView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(previous)))
+        content.addSubview(leftTapView)
+
+        NSLayoutConstraint.activate([
+            rightTapView.rightAnchor.constraint(equalTo: content.rightAnchor),
+            rightTapView.topAnchor.constraint(equalTo: buttonStack.bottomAnchor),
+            rightTapView.bottomAnchor.constraint(equalTo: content.bottomAnchor),
+            rightTapView.widthAnchor.constraint(equalTo: content.widthAnchor, multiplier: 0.33),
+        ])
+
+        NSLayoutConstraint.activate([
+            leftTapView.leftAnchor.constraint(equalTo: content.leftAnchor),
+            leftTapView.topAnchor.constraint(equalTo: buttonStack.bottomAnchor),
+            leftTapView.bottomAnchor.constraint(equalTo: content.bottomAnchor),
+            leftTapView.widthAnchor.constraint(equalTo: content.widthAnchor, multiplier: 0.33),
+        ])
+
         NSLayoutConstraint.activate([
             thumbsUp.bottomAnchor.constraint(equalTo: background.bottomAnchor, constant: -10),
             thumbsUp.centerXAnchor.constraint(equalTo: view.centerXAnchor),
@@ -159,10 +185,10 @@ class StoriesViewController: UIViewController {
         ])
 
         NSLayoutConstraint.activate([
-            segmentedProgress.leftAnchor.constraint(equalTo: content.leftAnchor, constant: 20),
-            segmentedProgress.rightAnchor.constraint(equalTo: buttonStack.leftAnchor, constant: -20),
-            segmentedProgress.heightAnchor.constraint(equalToConstant: 4),
-            segmentedProgress.centerYAnchor.constraint(equalTo: buttonStack.centerYAnchor),
+            progress.leftAnchor.constraint(equalTo: content.leftAnchor, constant: 20),
+            progress.rightAnchor.constraint(equalTo: buttonStack.leftAnchor, constant: -20),
+            progress.heightAnchor.constraint(equalToConstant: 4),
+            progress.centerYAnchor.constraint(equalTo: buttonStack.centerYAnchor),
         ])
 
         NSLayoutConstraint.activate([
@@ -196,13 +222,14 @@ class StoriesViewController: UIViewController {
             print("AVAudioSession error: \(error)")
         }
 
-        player.play()
-        segmentedProgress.startAnimation()
+        player.playTrack()
+        progress.startAnimation()
+        progress.isPaused = true
     }
 
     // @TODO allow deselecting reaction?
     @objc private func didReact(_ sender: UIButton) {
-        let item = player.currentItem()
+        let item = feed.stories[player.currentTrack]
 
         if feed.user.id == UserDefaults.standard.integer(forKey: UserDefaultsKeys.userId) {
             return
@@ -224,20 +251,20 @@ class StoriesViewController: UIViewController {
     }
 
     @objc private func exitTapped() {
-        player.stop()
+        player.pause()
         dismiss(animated: true)
     }
 
     @objc private func menuTapped() {
-        let item = player.currentItem()
+        let item = feed.stories[player.currentTrack]
 
         player.pause()
-        segmentedProgress.isPaused = true
+        progress.isPaused = true
 
         let menu = AlertController(title: nil, message: nil, preferredStyle: .actionSheet)
         menu.willDismissHandler = {
             self.player.unpause()
-            self.segmentedProgress.isPaused = false
+            self.progress.isPaused = false
         }
 
         let delete = UIAlertAction(title: NSLocalizedString("delete", comment: ""), style: .destructive, handler: { _ in
@@ -251,15 +278,38 @@ class StoriesViewController: UIViewController {
 
         present(menu, animated: true)
     }
+
+    @objc private func skip() {
+        progress.skip()
+        player.next()
+    }
+
+    @objc private func previous() {
+        progress.rewind()
+        player.previous()
+    }
 }
 
 extension StoriesViewController: StoryPlayerDelegate {
-    func didReachEnd() {
-        player.stop()
-        dismiss(animated: true)
+    func didStartBuffering(_: StoryPlayer) {
+        progress.isPaused = true
     }
 
-    func startedPlaying(story: APIClient.Story) {
+    func didEndBuffering(_: StoryPlayer) {
+        if !progress.isPaused {
+            return
+        }
+
+        progress.isPaused = false
+    }
+
+    func didStartPlaying(_: StoryPlayer, itemAt index: Int) {
+        let story = feed.stories[index]
+
+        if index != 0, progress.currentIndex < index {
+            progress.skip()
+        }
+
         posted.text = Date(timeIntervalSince1970: TimeInterval(story.deviceTimestamp)).timeAgoDisplay()
 
         thumbsUp.count = 0
@@ -279,10 +329,15 @@ extension StoriesViewController: StoryPlayerDelegate {
             }
         }
     }
+
+    func didReachEnd(_ player: StoryPlayer) {
+        player.pause()
+        dismiss(animated: true)
+    }
 }
 
 extension StoriesViewController: StoriesProgressBarDataSource {
     func storiesProgressBar(progressBar _: StoriesProgressBar, durationForItemAt index: Int) -> TimeInterval {
-        return TimeInterval(Float(CMTimeGetSeconds(player.playerItems[index].asset.duration)))
+        return player.duration(for: index)
     }
 }